+++ api/src/main/java/com/app/sip/model/Station.java
@@ -18,10 +18,10 @@ public class Station {
     private String address;
     private String city;
     private String postalCode;
-    private Long priceFuel95;
-    private Long priceFuel98;
-    private Long priceFuelDiesel;
-    private Long priceFuelLpg;
+    private Double priceFuel95;
+    private Double priceFuel98;
+    private Double priceFuelDiesel;
+    private Double priceFuelLpg;
 
     @OneToOne(cascade = CascadeType.PERSIST)
     private Brand brand;
